requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
